requirement.txt
torch
torchvision
torchaudio

transformers
datasets
accelerate

numpy<2
pandas
scikit-learn

matplotlib
seaborn

tqdm
jupyter
ipykernel

pillow
opencv-python
psutil

peft
bitsandbytes
evaluate
sentencepiece
einops

wandb